Refactor notification components
Consolidate legacy status components into a unified system with a shared
severity configuration. Introduces a `NotificationHost` for store-driven
global notifications (pill, card, banner variants) and moves ephemeral
toasts to an inline namespace.
