+++ src/shared/components/ui/notifications/index.ts
@@ -1,6 +1,27 @@
 // src/shared/components/ui/notifications/index.ts
-// Interchangeable notification components — all use notificationsStore
+//
+// Notification components split into two families:
+//
+// global/  — viewport-teleported, driven by notificationsStore
+//   NotificationHost   mount once in App.vue; `variant` prop selects the visual
+//   NotificationPill   small corner pill (default)
+//   NotificationCard   corner card with dismiss + progress bar
+//   NotificationBanner full-width edge bar with dismiss + progress bar
+//
+// inline/  — positioned relative to a parent, driven by local props
+//   InlineToast        dark pill for ephemeral confirmations ("Copied!")
 
-export { default as StatusBar } from './StatusBar.vue';
-export { default as StatusCorner } from './StatusCorner.vue';
-export { default as SubtleProgress } from './SubtleProgress.vue';
+// Global (store-driven, teleported to body)
+export {
+  NotificationHost,
+  NotificationPill,
+  NotificationCard,
+  NotificationBanner,
+} from './global';
+
+// Inline (prop-driven, parent-relative)
+export { InlineToast } from './inline';
+
+// Shared severity config
+export { getSeverityMeta, getInvertedColors, getStandardColors, getBannerColors } from './severityConfig';
+export type { SeverityMeta, SeverityColors } from './severityConfig';
