+++ src/shared/components/ui/notifications/inline/index.ts
@@ -1,0 +1,6 @@
+// src/shared/components/ui/notifications/inline/index.ts
+//
+// Inline (parent-relative, prop-driven) notification components.
+// No store coupling — consumers manage visibility with local refs.
+
+export { default as InlineToast } from './InlineToast.vue';
